Add support for #![no_std]
This allow to use Instant in no_std crate.
Added a Instant::from_duration that can be used to construct an instant
from a Duration when not using the Rust std lib
