+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.1.13
+
+## Added
+- Add enabled by default `std` feature which can be disabled to use this crate in `#![no_std]`  crates
+
 # v0.1.12
 ## Added 
 - Add `SystemTime` which works in both native and WASM environments.
